Изменент скрипт деплоя

+++ .bin/deploy.sh
@@ -1,51 +1,64 @@
-#! /bin/sh
+#!/usr/bin/env bash
 
-# -e - опция, требующая от bash немедленно выйти, если какая-либо команда имеет ненулевой
-# статус выхода
-# -u - опция приводит к немедленному завершению работы программы при установке ссылки
-# на любую неопределенную переменную (за исключением $* и $@)
 set -eu
 
-branch=$1
-commit=$2
+YELLOW_BG='\033[43m'
+YELLOW_FG='\033[43m'
+BLACK_FG='\033[30m'
+NC='\033[0m'
 
-echo "target branch: $branch"
-echo "commit sha: $commit"
+if [ ! $# -eq 0 ]
+then
+	branch=$1
+	commit=$2
 
-echo "=== Retrieving source code ($commit) ==="
+	echo -e "${YELLOW_FG}target branch: $branch${NC}"
+	echo -e "${YELLOW_FG}commit sha: $commit$branch${NC}"
 
-git reset --hard
-git clean -fd
-git checkout "$branch" -f
-git fetch origin "$branch"
-git checkout "$commit" -f
+	echo -e "\n${YELLOW_BG}${BLACK_FG}   Retrieving source code ($commit)   ${NC}"
+
+	git reset --hard
+	git clean -fd
+	git checkout "$branch" -f
+	git fetch origin "$branch"
+	git checkout "$commit" -f
+fi
+
+echo -e "\n${YELLOW_BG}${BLACK_FG}   Installing composer dependencies excluding DEV dependencies   ${NC}"
+./corn composer install --no-dev --no-interaction --prefer-dist --optimize-autoloader
+rm -f ./bootstrap/cache/packages.php
+rm -f ./bootstrap/cache/services.php
+./corn art package:discover
 
 # Создаем символьные ссылки если их нет
 if [ ! -h ./public/storage ]
 then
-	docker-compose exec -T php ./artisan storage:link
+	./corn artisan storage:link
 fi
 
-echo "=== Installing composer dependencies excluding DEV dependencies ==="
-docker-compose run composer install --no-dev --no-interaction --prefer-dist --optimize-autoloader --no-ansi;
+echo -e "\n${YELLOW_BG}${BLACK_FG}   Applying migrations   ${NC}"
+./corn artisan migrate --force
 
-echo "=== Applying migrations ==="
-docker-compose exec -T php ./artisan migrate --force
+echo -e "\n${YELLOW_BG}${BLACK_FG}   Data seeding   ${NC}"
+./corn artisan db:seed --force
 
-echo "=== Data seeding ==="
-docker-compose exec -T php ./artisan db:seed --force
+echo -e "\n${YELLOW_BG}${BLACK_FG}   Fresh cache   ${NC}"
+./corn artisan optimize -n
+./corn artisan event:cache -n
 
-echo "=== Fresh cache ==="
-docker-compose exec -T php ./artisan optimize -n
-docker-compose exec -T php ./artisan event:cache -n
+echo -e "\n${YELLOW_BG}${BLACK_FG}   Queue restarting   ${NC}"
+./corn artisan queue:restart
 
-echo "=== Queue restarting ==="
-docker-compose exec -T php ./artisan queue:restart
-docker-compose restart queue
+if [ -n `docker-compose ps -q nginx` ] && [ -n `docker ps -q --no-trunc | grep $(docker-compose ps -q nginx)` ]
+then
+	echo -e "\n${YELLOW_BG}${BLACK_FG}   NGINX config reloading   ${NC}"
+	./corn exec nginx nginx -s reload
+fi
 
-echo "=== NGINX config reloading ==="
-docker-compose exec -T nginx nginx -s reload
-
-echo "=== Nuxt reloading ==="
-docker-compose exec -T nodejs npm ci
-docker-compose restart nodejs
+echo -e "\n${YELLOW_BG}${BLACK_FG}   Frontend rebuilding   ${NC}"
+./corn npm install --no-audit
+./corn npm run generate
+if [ -n `docker-compose ps -q nodejs` ] && [ -n `docker ps -q --no-trunc | grep $(docker-compose ps -q nodejs)` ]
+then
+	./corn restart nodejs
+fi
